refactored in a new abstract class
